+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
   When switching the color style to one that does not support the 2025 specification, it will automatically switch to the 2021 specification.
 
+- Added a reset button to the right of all seed color pickers.
+
 ## 0.3.0
 
 Release Date: 2025-11-17